refactor(QA): moved services into dir for keep sorted
